requirements.txt
litellm
fastapi
uvicorn
pydantic
openai
sqlite3
pandas
openpyxl
python-magic
python-multipart
jinja2
requests
python-dotenv
